src/test/java/com/example/programmers/s67259Test.java
package com.example.programmers;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;

class s67259Test {

  @Test
  void solution() {
    assertEquals(900, s67259.solution(new int[][]{{0, 0, 0}, {0, 0, 0}, {0, 0, 0}}));
    assertEquals(3800, s67259.solution(
        new int[][]{{0, 0, 0, 0, 0, 0, 0, 1}, {0, 0, 0, 0, 0, 0, 0, 0}, {0, 0, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 1, 0, 0, 0}, {0, 0, 0, 1, 0, 0, 0, 1}, {0, 0, 1, 0, 0, 0, 1, 0},
            {0, 1, 0, 0, 0, 1, 0, 0}, {1, 0, 0, 0, 0, 0, 0, 0}}));
    assertEquals(2100,
        s67259.solution(new int[][]{{0, 0, 1, 0}, {0, 0, 0, 0}, {0, 1, 0, 1}, {1, 0, 0, 0}}));
    assertEquals(3200, s67259.solution(
        new int[][]{{0, 0, 0, 0, 0, 0}, {0, 1, 1, 1, 1, 0}, {0, 0, 1, 0, 0, 0}, {1, 0, 0, 1, 0, 1},
            {0, 1, 0, 0, 0, 1}, {0, 0, 0, 0, 0, 0}}));
  }
}
